requirements.txt
google-adk[a2a]
google-genai>=1.0.0
a2a-sdk>=0.3.22
pydantic>=2.10.0
python-dotenv
google-cloud-aiplatform
uvicorn
starlette
cloudpickle